+++ frontend/src/pages/Profile/index.ts
@@ -1,22 +1,11 @@
 import i18next from '@/config/i18n';
 import { Page } from '@/core/Page';
 import CommonLayout from '@/layouts/common/index';
+import { IBlockchainScore, ITournamentHistory } from '@/models/interface';
 import { checkUserAccess } from '@/models/User/auth';
 import { fetchCurrentUser } from '@/models/User/repository';
 import { setUserLanguage } from '@/utils/language';
 import { updateInnerHTML, updateText } from '@/utils/updateElements';
-
-interface ITournamentHistory {
-  date: string;
-  result: string;
-}
-
-// TODO: ブロックチェーンスコア機能の廃止
-// NOTE: バックエンドは削除済
-interface IBlockchainScore {
-  txHash: string;
-  score: number;
-}
 
 const languageNames: { [key: string]: string } = {
   en: 'English',
